# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "This Api for Photocontest App",
        "description": "Provide all the service for the Photocontest app",
        "termsOfService": "",
        "contact": {
            "email": "abdalrahman.alsharawy@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/items/{item}": {
            "get": {
                "tags": [
                    "items"
                ],
                "description": "Returns item.",
                "operationId": "show",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item object."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "items"
                ],
                "description": "create item.",
                "operationId": "update",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "available_stock",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "image_url",
                        "in": "formData",
                        "description": "Base64 string",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "items"
                ],
                "description": "Delete item.",
                "operationId": "destroy",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item object."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Returns List of users.",
                "operationId": "index",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Registeration.",
                "operationId": "store",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "full_name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "mobile_number",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/users/{user}": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Delete a user.",
                "operationId": "destroy",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "id of the user",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users List."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "description": "Registeration.",
                "operationId": "store",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "full_name",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_name",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mobile_number",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "shipping_address1",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "shipping_address2",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "shipping_country",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "shipping_city",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "description": "get user .",
                "operationId": "show",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "login.",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        }
    },
...
}